# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Notes App API",
    "description": "Backend APIs for a multi-user notes service.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/", "description": "Current server" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Note": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "content": { "type": "string" },
          "is_pinned": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      }
    }
  },
  "paths": {
    "/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } } },
          "400": { "description": "Bad request" },
          "409": { "description": "Email already exists" }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Authenticate user and get a JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid email or password" }
        }
      }
    },
    "/notes": {
      "get": {
        "summary": "Get all notes for the authenticated user (supports pagination)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "List of notes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Note" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create a new note",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string" },
                  "content": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Note created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Note" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/notes/{id}": {
      "get": {
        "summary": "Get a specific note by id",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Note data", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Note" } } } },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Note not found" }
        }
      },
      "put": {
        "summary": "Update an existing note",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string" },
                  "content": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated note", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Note" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Note not found" }
        }
      },
      "delete": {
        "summary": "Delete a note",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Note deleted" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Note not found" }
        }
      }
    },
    "/notes/{id}/share": {
      "post": {
        "summary": "Share a note with another user",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["share_with_email"],
                "properties": {
                  "share_with_email": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Shared successfully", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } } },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Note or user not found" }
        }
      }
    },
    "/notes/{id}/pin": {
      "patch": {
        "summary": "Pin a note (custom feature)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Note pinned", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Note" } } } },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Note not found" }
        }
      }
    },
    "/notes/{id}/unpin": {
      "patch": {
        "summary": "Unpin a note (custom feature)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Note unpinned", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Note" } } } },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Note not found" }
        }
      }
    },
    "/about": {
      "get": {
        "summary": "About this project",
        "responses": {
          "200": { "description": "About info" }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "summary": "Returns this OpenAPI spec",
        "responses": {
          "200": { "description": "OpenAPI JSON" }
        }
      }
    }
  }
}